fix: address Sentry backlog and harden persistence (#1475)
* fix(mister): keep unsupported-system errors out of Sentry

* fix(mister): keep missing requested files out of Sentry

* feat(api): add privacy-safe timeout diagnostics

* fix(mister): classify script-busy refusals as expected

* fix(config): classify unknown launcher fields as expected

* fix: harden transaction cleanup and catalog recovery

Discard buffered writes during rollback, preserve valid arcade catalogs, recognize the full SNES name, and classify local service refusals as warnings.

* fix(zapscript): classify unsupported control actions as expected

* fix(mediadb): preserve optimization progress on cancellation

* fix: handle optional scanner absence within subsystem boundaries

* fix(config): preserve existing files with atomic saves

* fix(mister): preserve startup scripts during load and save

* fix(mister): replace AmigaCD32 configuration atomically

* fix(mister): save profile labels atomically

* fix(retroarch): atomically replace generated configuration

* fix(mediadb): diagnose indexing state lock failures

* fix(database): stop media recreation when sidecar cleanup fails

* fix(mediadb): mark corruption returned by systems queries

* fix(userdb): mark corruption in media history operations

* fix(mediascanner): quiet secondary corruption status failures

* fix(userdb): upsert categorized inbox messages atomically

* fix(api): classify typed validation errors as expected

* fix(api): warn on disconnected HTTP response clients

* fix(api): reject non-WebSocket upgrade requests early

* fix(api): include safe status diagnostics for failed handshakes

* fix(api): stop systems fallback after context cancellation

* fix(readers): classify absent writable readers as expected

* fix(api): avoid duplicate reader write error reports

* fix(tty2oled): avoid signed overflow in picture selection

* fix(tty2oled): accept nil media as a clear request

* fix(launchers): diagnose missing executable and working directory

* fix(mister): preserve script stat errors and classify missing files

* fix(mister): treat empty core names as transient rewrites

* fix(windows): warn on duplicate service launches

* fix(tui): preserve original error when fallback fails

* fix(systray): warn on duplicate config reload failures

* fix(remote): quietly stop after acceptance transition loses race

* fix(zapscript): classify invalid advanced arguments as expected

* fix(zapscript): reject entirely invalid random system lists

* fix(zapscript): omit command text and stderr from execute errors

* fix(kodi): bound responses and report safe protocol diagnostics

* fix(arcadedb): omit GitHub response bodies from failures

+++ pkg/readers/tty2oled/mappings.go
@@ -154,7 +154,8 @@ func selectPictureVariant(baseName string) string {
 		log.Error().Str("baseName", baseName).Msg("no variants available")
 		return baseName // fallback to original name
 	}
-	selected := int(hash) % variantsLen
+	// Reduce before converting: uint32 hashes can overflow int on 32-bit targets.
+	selected := int(uint(hash) % uint(variantsLen))
 
 	// Safety check to prevent index out of bounds
 	if selected < 0 || selected >= len(variants) {
